flask_api_for_rec/requirements.txt
flask-restful
numpy
pandas
pymongo[srv]
scikit-learn
flask-cors